e2e/requirements.txt
pytest==8.3.4
playwright==1.50.0
PyYAML==6.0.2
python-dotenv==1.0.1
